Fix-interfece/singltone/lambda/ctrl+alt+L
