+++ src/main/java/com/example/moviereviewplatform/dao/AbstractHibernateDao.java
@@ -4,7 +4,6 @@ import com.example.moviereviewplatform.util.HibernateUtil;
 import org.hibernate.Session;
 import org.hibernate.SessionFactory;
 import org.hibernate.Transaction;
-
 import java.io.Serializable;
 import java.util.List;
 import java.util.Optional;
@@ -33,6 +32,7 @@ public abstract class AbstractHibernateDao<K, T> implements Dao<K, T> {
             throw new RuntimeException("Ошибка при сохранении объекта", e);
         }
     }
+
     public List<T> findAll() {
         try (Session session = getSession()) {
             return session.createQuery("FROM " + entityClass.getName(), entityClass).getResultList();
@@ -40,6 +40,7 @@ public abstract class AbstractHibernateDao<K, T> implements Dao<K, T> {
             throw new RuntimeException("Ошибка при выполнении запроса findAll для " + entityClass.getName(), e);
         }
     }
+
     public Optional<T> findById(K id) {
         try (Session session = getSession()) {
             T entity = session.get(entityClass, (Serializable) id);
@@ -48,6 +49,7 @@ public abstract class AbstractHibernateDao<K, T> implements Dao<K, T> {
             throw new RuntimeException("Ошибка при поиске объекта по id: " + id, e);
         }
     }
+
     public boolean deleteById(K id) {
         Transaction transaction = null;
         try (Session session = getSession()) {
