+++ src/main/java/com/example/moviereviewplatform/dao/Dao.java
@@ -1,15 +1,20 @@
 package com.example.moviereviewplatform.dao;
 
-import com.example.moviereviewplatform.entity.Movies;
-
 import java.util.List;
 import java.util.Optional;
 
+/**
+ * @param <K>
+ * @param <T>
+ */
 public interface Dao<K, T> {
 
     List<T> findAll();
+
     Optional<T> findById(K id);
+
     void save(T entity);
+
     boolean deleteById(K id);
