+++ src/main/java/com/example/moviereviewplatform/dao/MovieDao.java
@@ -1,8 +1,8 @@
 package com.example.moviereviewplatform.dao;
 
 import com.example.moviereviewplatform.entity.Movies;
-
 import java.util.List;
+
 
 public interface MovieDao extends Dao<Integer, Movies> {
     List<Movies> getTopRatedMovies();
