+++ src/main/java/com/example/moviereviewplatform/dao/ReviewDao.java
@@ -1,7 +1,6 @@
 package com.example.moviereviewplatform.dao;
 
 import com.example.moviereviewplatform.entity.Reviews;
-
 import java.util.List;
 
 public interface ReviewDao extends Dao<Integer, Reviews> {
