+++ src/main/java/com/example/moviereviewplatform/dao/UserDao.java
@@ -1,7 +1,6 @@
 package com.example.moviereviewplatform.dao;
 
 import com.example.moviereviewplatform.entity.User;
-
 import java.util.Optional;
 
 public interface UserDao extends Dao<Integer, User> {
