+++ src/main/java/com/example/moviereviewplatform/dao/impl/MovieDaoImpl.java
@@ -1,6 +1,7 @@
-package com.example.moviereviewplatform.dao;
+package com.example.moviereviewplatform.dao.impl;
 
-import com.example.moviereviewplatform.dto.MovieDto;
+import com.example.moviereviewplatform.dao.AbstractHibernateDao;
+import com.example.moviereviewplatform.dao.MovieDao;
 import com.example.moviereviewplatform.entity.Movies;
 import com.example.moviereviewplatform.util.HibernateUtil;
 import org.hibernate.Query;
@@ -9,7 +10,7 @@ import org.hibernate.SessionFactory;
 
 import java.util.List;
 
-public class MovieDaoImpl extends AbstractHibernateDao<Integer, Movies> implements MovieDao{
+public class MovieDaoImpl extends AbstractHibernateDao<Integer, Movies> implements MovieDao {
 
     private static final MovieDaoImpl INSTANCE = new MovieDaoImpl(HibernateUtil.getSessionFactory());
 
@@ -33,23 +34,15 @@ public class MovieDaoImpl extends AbstractHibernateDao<Integer, Movies> implemen
             throw new RuntimeException("Ошибка при получении топ фильмов", e);
         }
     }
-    public List<MovieDto> findAllPaginated(int page, int size) {
+
+    public List<Movies> findAllPaginated(int page, int size) {
         try (Session session = HibernateUtil.getSessionFactory().openSession()) {
             Query<Movies> query = session.createQuery("FROM Movies", Movies.class);
             query.setFirstResult((page - 1) * size); // Устанавливаем смещение
             query.setMaxResults(size); // Устанавливаем лимит
 
             // Преобразование списка сущностей Movies в DTO
-            return query.list().stream()
-                    .map(movie -> new MovieDto(
-                            movie.getId(),
-                            movie.getName(),
-                            movie.getGenre(),
-                            movie.getDescription(),
-                            movie.getPoster_url(),
-                            movie.getRelease_date()
-                    ))
-                    .toList();
+            return query.list();
         } catch (Exception e) {
             throw new RuntimeException("Ошибка при получении фильмов с пагинацией", e);
         }
