+++ src/main/java/com/example/moviereviewplatform/dto/CreateMovieDto.java
@@ -3,8 +3,8 @@ package com.example.moviereviewplatform.dto;
 import jakarta.servlet.http.Part;
 import lombok.Builder;
 import lombok.Value;
-
 import java.time.LocalDate;
+
 @Value
 @Builder
 public class CreateMovieDto {
