+++ src/main/java/com/example/moviereviewplatform/dto/WatchlistDto.java
@@ -1,8 +1,6 @@
 package com.example.moviereviewplatform.dto;
 
-import lombok.AllArgsConstructor;
 import lombok.Builder;
-import lombok.NoArgsConstructor;
 import lombok.Value;
 
 @Value
