+++ src/main/java/com/example/moviereviewplatform/entity/Movies.java
@@ -7,7 +7,6 @@ import lombok.Data;
 import lombok.NoArgsConstructor;
 
 
-
 import javax.persistence.*;
 import java.time.LocalDate;
 import java.util.List;
