+++ src/main/java/com/example/moviereviewplatform/entity/Role.java
@@ -7,9 +7,9 @@ public enum Role {
     USER,
     ADMIN;
 
-        public static Optional<Role> find(String role){
-            return Arrays.stream(values())
-                    .filter(it -> it.name().equals(role))
-                    .findFirst();
-        }
+    public static Optional<Role> find(String role) {
+        return Arrays.stream(values())
+                .filter(it -> it.name().equals(role))
+                .findFirst();
+    }
 }
