+++ src/main/java/com/example/moviereviewplatform/filter/AuthorizationFilter.java
@@ -16,6 +16,7 @@ import static com.example.moviereviewplatform.util.UrlPath.*;
 public class AuthorizationFilter implements Filter {
 
     private static final Set<String> PUBLIC_PATH = Set.of(LOGIN, REGISTRATION, MOVIES, REVIEW, LOCATE);
+
     @Override
     public void doFilter(ServletRequest servletRequest, ServletResponse servletResponse, FilterChain filterChain) throws IOException, ServletException {
