+++ src/main/java/com/example/moviereviewplatform/mapper/Mapper.java
@@ -1,9 +1,6 @@
 package com.example.moviereviewplatform.mapper;
 
-import com.example.moviereviewplatform.dto.CreateMovieDto;
-import com.example.moviereviewplatform.entity.Movies;
-
-public interface Mapper <F, T>{
+public interface Mapper<F, T> {
 
 
     T mapFrom(F object);
