+++ src/main/java/com/example/moviereviewplatform/mapper/MovieMapper.java
@@ -1,15 +1,13 @@
 package com.example.moviereviewplatform.mapper;
 
 import com.example.moviereviewplatform.dto.MovieDto;
-import com.example.moviereviewplatform.dto.UserDto;
 import com.example.moviereviewplatform.entity.Movies;
-import com.example.moviereviewplatform.entity.User;
 import lombok.NoArgsConstructor;
 
 import static lombok.AccessLevel.PRIVATE;
 
 @NoArgsConstructor(access = PRIVATE)
-public class MovieMapper implements Mapper<Movies, MovieDto>{
+public class MovieMapper implements Mapper<Movies, MovieDto> {
 
     private static final MovieMapper INSTANCE = new MovieMapper();
