+++ src/main/java/com/example/moviereviewplatform/servlet/UserController.java
@@ -1,26 +1,23 @@
 package com.example.moviereviewplatform.servlet;
 
 import com.example.moviereviewplatform.dto.UserDto;
-import com.example.moviereviewplatform.entity.User;
-import com.example.moviereviewplatform.service.ImageService;
-import com.example.moviereviewplatform.service.ReviewService;
-import com.example.moviereviewplatform.service.UserService;
-import com.example.moviereviewplatform.service.WatchlistService;
-import com.example.moviereviewplatform.util.PropertiesUtil;
+import com.example.moviereviewplatform.service.impl.ImageServiceImpl;
+import com.example.moviereviewplatform.service.impl.UserServiceImpl;
+import com.example.moviereviewplatform.service.impl.WatchlistServiceImpl;
 import jakarta.servlet.ServletException;
 import jakarta.servlet.annotation.MultipartConfig;
 import jakarta.servlet.annotation.WebServlet;
 import jakarta.servlet.http.*;
 
 import java.io.IOException;
-import java.util.Optional;
+
 @MultipartConfig
 @WebServlet("/user")
 public class UserController extends HttpServlet {
     private static final String USER = "user";
-    private final UserService userService = UserService.getInstance();
-    private final ImageService imageService = ImageService.getInstance();
-    private final WatchlistService watchlistService = WatchlistService.getInstance();
+    private final UserServiceImpl userServiceImpl = UserServiceImpl.getInstance();
+    private final ImageServiceImpl imageServiceImpl = ImageServiceImpl.getInstance();
+    private final WatchlistServiceImpl watchlistServiceImpl = WatchlistServiceImpl.getInstance();
 
     protected void doGet(HttpServletRequest req, HttpServletResponse resp) throws ServletException, IOException {
         // Получаем сессию и пользователя из неё
@@ -31,13 +28,13 @@ public class UserController extends HttpServlet {
             Integer userId = user.getId();
 
             // Устанавливаем атрибуты для изображения и информации о пользователе
-            req.setAttribute("image", imageService.get(user.getImage()));
+            req.setAttribute("image", imageServiceImpl.get(user.getImage()));
             req.setAttribute("user", user);
             req.setAttribute("userId", userId);
 
             // Получаем списки watching и watched
-            var watchingList = watchlistService.getUserWatchlistByType(userId, "watching");
-            var watchedList = watchlistService.getUserWatchlistByType(userId, "watched");
+            var watchingList = watchlistServiceImpl.getUserWatchlistByType(userId, "watching");
+            var watchedList = watchlistServiceImpl.getUserWatchlistByType(userId, "watched");
 
             // Устанавливаем атрибуты для списка
             req.setAttribute("watchingList", watchingList);
