+++ src/main/java/com/example/moviereviewplatform/util/HibernateUtil.java
@@ -8,7 +8,6 @@ public class HibernateUtil {
     private static SessionFactory sessionFactory;
 
     private HibernateUtil() {
-
     }
 
     public static SessionFactory getSessionFactory() {
