+++ src/main/java/com/example/moviereviewplatform/util/JspHelper.java
@@ -2,7 +2,6 @@ package com.example.moviereviewplatform.util;
 
 public class JspHelper {
     private static final String JSP_FORMAT = "/WEB-INF/jsp/%s.jsp";
-
     public static String getPath(String jspName) {
         return String.format(JSP_FORMAT, jspName);
     }
