+++ src/main/java/com/example/moviereviewplatform/util/LocalDateTimeConverter.java
@@ -1,7 +1,6 @@
 package com.example.moviereviewplatform.util;
 
 
-
 import javax.persistence.AttributeConverter;
 import javax.persistence.Converter;
 import java.sql.Timestamp;
@@ -10,13 +9,12 @@ import java.time.LocalDateTime;
 @Converter
 public class LocalDateTimeConverter implements AttributeConverter<LocalDateTime, Timestamp> {
 
-@Override
-public Timestamp convertToDatabaseColumn(LocalDateTime attribute) {
+    @Override
+    public Timestamp convertToDatabaseColumn(LocalDateTime attribute) {
         return attribute == null ? null : Timestamp.valueOf(attribute);
-        }
-@Override
-
-public LocalDateTime convertToEntityAttribute(Timestamp dbData) {
+    }
+    @Override
+    public LocalDateTime convertToEntityAttribute(Timestamp dbData) {
         return dbData == null ? null : dbData.toLocalDateTime();
-        }
-        }
+    }
+}
