+++ src/main/java/com/example/moviereviewplatform/util/PropertiesUtil.java
@@ -1,7 +1,5 @@
 package com.example.moviereviewplatform.util;
 
-import lombok.experimental.UtilityClass;
-
 import java.io.IOException;
 import java.io.InputStream;
 import java.util.Properties;
